Names search deals with duplicates
Occasionally, taxa in different nomenclature systems have the same name. For example, Iris is both a plant and an insect. You can now give the PIA an expected phylogenetic range into which your reads should fall. This gives the PIA more information to distinguish between these names. If there are still duplicates inside the range, it takes the one with the highest rank, and if there are multiple in the highest rank, it chooses randomly. But it's a lot less random choosing than before.

To output this extra information, the .intersects and Summary_Basic.txt outputs now contain taxonomic ID as well as taxon name.