Assigns BLAST hits to intersection of associated taxa
BLAST hits are to one reference sequence, but that sequence can be associated with identical sequence in different taxa. Like in MEGAN, we assume that this means the sequence is conservative and we assign it to the intersection of the associated taxa instead of the "main" taxon it came with.